# swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: Collecto API
  description: API de autenticación y gestión de sesión para la plataforma Collecto
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/collecto/Collecto/1.0.0
  - url: http://localhost:3000
    description: Servidor de desarrollo local
paths:
...
  /api/users/{username}/listings:
    get:
      tags:
        - Anuncios
      summary: Ver anuncios de un usuario
      description: Obtiene los anuncios de un usuario específico. Si el usuario está autenticado, incluye el estado de favoritos en los anuncios
      operationId: getUserAdverts
      parameters:
        - name: username
          in: path
          required: true
          description: Nombre de usuario del cual obtener los anuncios
          schema:
            type: string
      responses:
        '200':
          description: Lista de anuncios del usuario con estado de favoritos si el usuario está autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    description: Total de anuncios del usuario
                  adverts:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: ID del anuncio
                        title:
                          type: string
                          description: Título del anuncio
                        description:
                          type: string
                          description: Descripción del anuncio
                        price:
                          type: number
                          format: double
                          description: Precio del artículo
                        isFavorite:
                          type: boolean
                          description: Indica si el anuncio es favorito del usuario autenticado
        '404':
          description: Usuario no encontrado o sin anuncios
        '500':
          description: Error interno en el servidor
...
  /api/users/me:
    get:
      tags:
        - Usuarios
      summary: Obtener datos del propio usuario
      description: Devuelve los datos del usuario autenticado
      operationId: getCurrentUser
      responses:
        '200':
          description: Datos del usuario
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                  email:
                    type: string
                  firstName:
                    type: string
                  lastName:
                    type: string
                  phone:
                    type: string
                  location:
                    type: string
                  avatarUrl:
                    type: string
                  bio:
                    type: string
        '404':
          description: Usuario no encontrado
  /api/users/mee:
    put:
      tags:
        - Usuarios
      summary: Editar perfil del usuario
      description: Permite editar los datos del usuario autenticado
      operationId: editUserProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                phone:
                  type: string
                location:
                  type: string
                avatarUrl:
                  type: string
                bio:
                  type: string
      responses:
        '200':
          description: Perfil actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Perfil actualizado"
                  user:
                    $ref: '#/components/schemas/User'
        '404':
          description: Usuario no encontrado
  /api/userse/me:
    delete:
      tags:
        - Usuarios
      summary: Eliminar perfil del usuario
      description: Elimina el perfil del usuario autenticado
      operationId: deleteUserProfile
      responses:
        '200':
          description: Cuenta eliminada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Cuenta eliminada"
        '404':
          description: Usuario no encontrado
  /api/users/me/adverts:
    get:
      tags:
        - Anuncios
      summary: Ver anuncios de uno mismo
      description: Obtiene los anuncios del usuario autenticado. Incluye el total de anuncios publicados.
      operationId: getOwnAdverts
      responses:
        '200':
          description: Lista de anuncios del usuario autenticado, incluyendo el total de anuncios
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalAdverts:
                    type: integer
                    description: Total de anuncios publicados por el usuario
                  adverts:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: ID del anuncio
                        title:
                          type: string
                          description: Título del anuncio
                        description:
                          type: string
                          description: Descripción del anuncio
                        price:
                          type: number
                          format: double
                          description: Precio del artículo
        '404':
          description: Usuario no encontrado o sin anuncios
        '500':
          description: Error interno en el servidor
...
  /api/users/me/favorites:
    get:
      tags:
        - Favoritos
      summary: Obtener "Mis anuncios favoritos"
      description: Obtiene los anuncios favoritos del usuario autenticado. Incluye el total de favoritos.
      operationId: getUserFavorites
      responses:
        '200':
          description: Lista de anuncios favoritos del usuario, incluyendo el total de favoritos
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalFavorites:
                    type: integer
                    description: Total de anuncios favoritos del usuario
                  favorites:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: ID del anuncio favorito
                        title:
                          type: string
                          description: Título del anuncio favorito
                        description:
                          type: string
                          description: Descripción del anuncio favorito
                        price:
                          type: number
                          format: double
                          description: Precio del artículo favorito
        '404':
          description: Usuario no encontrado o sin favoritos
        '500':
          description: Error interno en el servidor
  /api/users/me/favorites/{listingId}:
    post:
      tags:
        - Favoritos
      summary: Agregar un anuncio a favoritos
      description: Permite agregar un anuncio a la lista de favoritos del usuario
      operationId: addFavorite
      parameters:
        - name: listingId
          in: path
          required: true
          description: ID del anuncio a agregar a favoritos
          schema:
            type: string
      responses:
        '201':
          description: Anuncio agregado a favoritos
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Añadido a favoritos"
        '404':
          description: Anuncio no encontrado
        '400':
          description: El anuncio ya está en tus favoritos
  /api/users/me/favoirtes/{listingId}:
    delete:
      tags:
        - Favoritos
      summary: Eliminar un anuncio de favoritos
      description: Permite eliminar un anuncio de la lista de favoritos del usuario
      operationId: removeFavorite
      parameters:
        - name: listingId
          in: path
          required: true
          description: ID del anuncio a eliminar de favoritos
          schema:
            type: string
      responses:
        '200':
          description: Anuncio eliminado de favoritos
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Eliminado de favoritos"
        '404':
          description: Anuncio no encontrado
        '400':
          description: El anuncio no está en tus favoritos
  /api/users/me/notifications:
    get:
      tags:
        - Notificaciones
      summary: Obtener "Mis notificaciones"
      description: Devuelve todas las notificaciones del usuario autenticado
      operationId: getUserNotifications
      responses:
        '200':
          description: Lista de notificaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '404':
          description: No tienes notificaciones
  /api/users/me/notifications/{id}/read:
    patch:
      tags:
        - Notificaciones
      summary: Marcar notificación como leída
      description: Permite marcar una notificación como leída
      operationId: markNotificationAsRead
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la notificación a marcar como leída
          schema:
            type: string
      responses:
        '200':
          description: Notificación marcada como leída
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Leída"
        '404':
          description: Notificación no encontrada
        '403':
          description: No tienes permiso para modificar esta notificación
  /api/users/me/notifications/favorite-status-change:
    post:
      tags:
        - Notificaciones
      summary: Notificación de cambio de estado de favorito
      description: Crea una notificación cuando el estado de un anuncio favorito cambia (vendido/reservado/disponible)
      operationId: notifyFavoriteStatusChange
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                advertId:
                  type: string
                status:
                  type: string
                  enum: ['sold', 'reserved', 'available']
      responses:
        '201':
          description: Notificación de cambio de estado de favorito creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Notificación de cambio de estado de favorito creada"
        '400':
          description: Estado inválido
  /api/users/me/notifications/favorite-price-change:
    post:
      tags:
        - Notificaciones
      summary: Notificación de cambio de precio de favorito
      description: Crea una notificación cuando el precio de un anuncio favorito cambia
      operationId: notifyPriceChange
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                advertId:
                  type: string
      responses:
        '201':
          description: Notificación de cambio de precio enviada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Notificación de cambio de precio enviada"
        '404':
          description: Anuncio no encontrado
  /api/users/me/notifications/favorite-removed:
    post:
      tags:
        - Notificaciones
      summary: Notificación de eliminación de favorito
      description: Crea una notificación cuando un anuncio es eliminado de favoritos
      operationId: notifyFavoriteRemoved
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                advertId:
                  type: string
      responses:
        '201':
          description: Notificación de eliminación de favorito enviada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Notificación de eliminación de favorito enviada"
        '404':
          description: Anuncio no encontrado
  /api/users/me/notifications/new-chat-message:
    post:
      tags:
        - Notificaciones
      summary: Notificación de nuevo mensaje en el chat
      description: Crea una notificación cuando un nuevo mensaje es enviado en un chat
      operationId: notifyNewChatMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chatId:
                  type: string
      responses:
        '201':
          description: Notificación de nuevo mensaje enviada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Notificación de nuevo mensaje enviada"
        '404':
          description: Chat no encontrado
  /api/users/me/chat/{listingId}:
    post:
      tags:
        - Chat
      summary: Crear un chat relacionado con un anuncio
      description: Crea un nuevo chat entre el usuario autenticado y el vendedor del anuncio
      operationId: createChat
      parameters:
        - name: listingId
          in: path
          required: true
          description: ID del anuncio relacionado con el chat
          schema:
            type: string
      responses:
        '201':
          description: Chat creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Chat creado exitosamente"
                  chatId:
                    type: string
        '404':
          description: Anuncio no encontrado
        '400':
          description: No puedes chatear contigo mismo
  /api/users/me/chat/{chatId}:
    post:
      tags:
        - Chat
      summary: Enviar un mensaje en un chat
      description: Permite al usuario autenticado enviar un mensaje en un chat existente
      operationId: sendMessageToChat
      parameters:
        - name: chatId
          in: path
          required: true
          description: ID del chat en el que se enviará el mensaje
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  description: Contenido del mensaje
      responses:
        '201':
          description: Mensaje enviado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Mensaje enviado"
        '404':
          description: Chat no encontrado
        '403':
          description: No tienes permiso para enviar mensajes en este chat
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          description: ID del usuario
        username:
          type: string
          description: Nombre de usuario
        email:
          type: string
          description: Correo electrónico
        firstName:
          type: string
          description: Primer nombre del usuario
        lastName:
          type: string
          description: Apellido del usuario
        phone:
          type: string
          description: Teléfono del usuario
        location:
          type: string
          description: Ubicación del usuario
        avatarUrl:
          type: string
          description: URL de la imagen de perfil del usuario
        bio:
          type: string
          description: Biografía del usuario
...
    Notification:
      type: object
      properties:
        user:
          type: string
          description: ID del usuario al que se le envió la notificación
        notificationType:
          type: string
          description: Tipo de notificación
        advertId:
          type: string
          description: ID del anuncio relacionado con la notificación
        message:
          type: string
          description: Mensaje de la notificación
        isRead:
          type: boolean
          description: Estado de lectura de la notificación
        createdAt:
          type: string
          format: date-time
          description: Fecha en que se creó la notificación
...